project/wikiproject/requirements.txt
constants
south
django_notify
